--- Logging/Controllers/View Controllers/Form 781/SideMenuTableViewController.swift
@@ -1,60 +1,0 @@
-//
-//  SideMenuTableViewController.swift
-//  Logging
-//
-//  Created by Bethany Morris on 10/21/20.
-//  Copyright © 2020 Christian Brechbuhl. All rights reserved.
-//
-
-import UIKit
-
-protocol SideMenuDelegate: class {
-    func toggleSideMenu()
-}
-
-class SideMenuTableViewController: UITableViewController {
-    
-    weak var delegate: SideMenuDelegate?
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-
-    @IBAction func menuButtonTapped(_ sender: UIButton) {
-        print("tapped")
-        delegate?.toggleSideMenu()
-    }
-
-    // MARK: - Table view data source
-
-//    override func numberOfSections(in tableView: UITableView) -> Int {
-//        // #warning Incomplete implementation, return the number of sections
-//        return 0
-//    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return 0
-    }
-
-    /*
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "reuseIdentifier", for: indexPath)
-
-        // Configure the cell...
-
-        return cell
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-} //End
